Put for= parameters at front of querystring
Acropolis puts the for= parameters at the front of the querystring,
so ensure that RES Search sends them in the same order as they
come back from Acropolis (to match the OSD template).

This is important, as we attempt to resolve the URI against the
context. If we're using a URI in a different format or order
from the one which Acropolis sends back in the search results,
we won't necessarily be able to find it in the RDF.