urldecode URI when retrieving from LOD
Search URIs including for= parameters should be urlencoded
when fetched (as they contain "http://" strings as the values
for the for= params).

However, the returned RDF contains a search dataset resource
whose URI is the urldecoded equivalent (i.e. it contains
"for=http://" strings). This was causing issues when
LOD did a resolve() on the URI, as the fetched URI and the
URI in the response RDF didn't marry up.

To enable LOD to locate the search result resource, first
do a fetch of the RDF; then urldecode the fetched URI and use
that as the parameter to locate().