Only URL-encode '#' characters in audience URIs
When an audience URI is sent to Acropolis in a search, the
search response's main dataset will have that URI with the
'#' characters replaced with '%23' (i.e. only '#' is URL-encoded
in the response).

Previously, the whole audience URI was URL-encoded, so there
was a mismatch when the RES search service tried to find the
search result resource.

Modify the URL-encoding of search parameters so that only the
'#' character is URL-encoded.