Allow endpoints to be full or relative URLs
The current arrangement for setting the location of API
endpoints on the search service is to set paths in the
index.php file. The host is then appended to the path when
calling the API from the web UI.

Instead, don't add the host to the path when determining the
API endpoint in the web UI. This means a full URL or just a
path can be used to set the endpoint location. In turn,
this opens up the possibility of using full URLs when setting
the API endpoints in the context of the Moodle plugin. These
can be generated using the moodle_url() function, which means
they should be correct even if Moodle is installed in a
sub-directory.