Add media type guess heuristic based on player URI
In some situations, we don't have information about the type
of the playable media (e.g. the BBC Teach resources which
point at YouTube don't have any indication that the YouTube
resources are videos).

Add a heuristic to guess the content type of a piece of media
based on the URI.